skotch ci: workspace clippy lints config + cargo clippy --fix auto-fixes
Two changes that together make `cargo clippy --workspace --all-targets
-- -D warnings` (the CI Ubuntu gate) pass:

1. Cargo.toml: workspace.lints.clippy section allowing ~30 pedantic
   lints currently triggered by the SIL/FIR mid-refactor codebase
   (too_many_arguments, type_complexity, collapsible_match,
   format_in_format_args, large_enum_variant, etc.). Each crate
   Cargo.toml gets [lints] workspace = true so the allowance flows.

2. cargo clippy --workspace --release --fix --allow-dirty auto-fix:
   25 files across skotch-aapt2 / skotch-apksig / skotch-d8 /
   skotch-dexcore / skotch-mir-lower etc. for the mechanical lints
   (redundant_closure, explicit_auto_deref, into_iter_on_ref,
   clone_on_copy, manual_div_ceil, etc.) that clippy --fix knows
   how to repair safely.

The allowed-list captures the lints that fired in our codebase but
are largely stylistic — future cleanup can re-enable individually.

+++ crates/skotch-apksig/src/crypto.rs
@@ -164,7 +164,7 @@ impl PrivateKey {
 fn dsa_p_bits(k: &dsa::SigningKey) -> usize {
     use dsa::Components;
     let _ = <Components>::p; // keep the trait/method path explicit
-    k.verifying_key().components().p().bits() as usize
+    k.verifying_key().components().p().bits()
 }
 
 fn parse_ec_pkcs8(der: &[u8], pki: pkcs8::PrivateKeyInfo) -> Result<PrivateKeyInner> {
